Renamed engine-iso-uploader in ovirt-iso-uploader
The package should have the same name of the application.
In this case, the package name seems better than the application
name, so I've renamed the application instead of the package.

engine-iso-uploader and a placeholder for its man page are kept
for backward compatibility until all the documentation will be
updated.

Change-Id: I5f6cc47a3e068a62d6e8bdcd20f360e1d2a53c1c
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>