upload: nfs: handle rename errors
Uploading to a NFS destination, if an error
were encountered while renaming the used temporary
file to the final destination,
the application displayed: uploaded successfully.

Now the application exits with exit status 3
(upload error) and display an error message.
Added also a debug log for recording the rename
action in the verbose log.

Change-Id: I8fc993d0fdc36924cbf3dd7ed50569ccdcfc75b5
Bug-Url: https://bugzilla.redhat.com/1005100
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>