UX: remove alarming warning
After upgrading from 3.1 users was warned about a legacy config
still present in conf files if the user forgot to update .conf file
with the one provided by the package as .rpmnew.
The tool is able to use the legacy configuration so no need to
alarm the user about that.

Change-Id: I6ad7695a30f133fd1171c8dae8977ef92ea65be7
Bug-Url: https://bugzilla.redhat.com/1031092
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>