error handling: handle invalid host
previously if an invalid host was specified an error like

ERROR: 'NoneType' object is not iterable

was printed.
Now a more explanatory error is showed:

ERROR: Unable to connect to REST API.  Message: The host name
"localhost" contained in the URL doesn't match any of the names in the
server certificate.
ERROR: Unable to get ISO domain data

Change-Id: I86ff1d2bcff923c76176554ae73beb47caa3d633
Bug-Url: https://bugzilla.redhat.com/1110756
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>